refactor: extract LargeFileInterceptor service from LcmContextEngine
Phase 2 of the engine.ts decomposition (#867). The ~950-line ingest-time
interception cluster (oversized <file> blocks, inline/native/base64
images, oversized tool results, generic raw payloads) moves to
src/large-file-interceptor.ts. The interceptor takes config, the summary
store, and a resolver callback for the optional model-backed large-file
summarizer, which stays in the engine's summarizer-resolution cluster.
Method bodies unchanged; the six ingestSingle call sites now go through
this.largeFileInterceptor.

Full suite: 1279 tests passing.
